+++ examples/inputs-simple-sqlite/src/schema/__generated__/User/index.ts
@@ -14,3 +14,29 @@ export {
   UserFollowersFieldObject,
   UserFollowingFieldObject
 } from './object.base';
+export {
+  createManyUserMutation,
+  createOneUserMutation,
+  deleteManyUserMutation,
+  deleteOneUserMutation,
+  updateManyUserMutation,
+  updateOneUserMutation,
+  upsertOneUserMutation,
+  createManyUserMutationObject,
+  createOneUserMutationObject,
+  deleteManyUserMutationObject,
+  deleteOneUserMutationObject,
+  updateManyUserMutationObject,
+  updateOneUserMutationObject,
+  upsertOneUserMutationObject
+} from './mutations';
+export {
+  findFirstUserQuery,
+  findManyUserQuery,
+  countUserQuery,
+  findUniqueUserQuery,
+  findFirstUserQueryObject,
+  findManyUserQueryObject,
+  countUserQueryObject,
+  findUniqueUserQueryObject
+} from './queries';