+++ examples/inputs-simple-sqlite/src/schema/__generated__/User/mutations/index.ts
@@ -1,0 +1,7 @@
+export { createManyUserMutation, createManyUserMutationObject } from './createMany.base';
+export { createOneUserMutation, createOneUserMutationObject } from './createOne.base';
+export { deleteManyUserMutation, deleteManyUserMutationObject } from './deleteMany.base';
+export { deleteOneUserMutation, deleteOneUserMutationObject } from './deleteOne.base';
+export { updateManyUserMutation, updateManyUserMutationObject } from './updateMany.base';
+export { updateOneUserMutation, updateOneUserMutationObject } from './updateOne.base';
+export { upsertOneUserMutation, upsertOneUserMutationObject } from './upsertOne.base';